chore: add core crate
