+++ Cargo.toml
@@ -1,16 +1,2 @@
-[package]
-name = "notion-to-markdown"
-version = "0.1.0"
-edition = "2021"
-
-
-[dependencies]
-comrak = "0.39.0"
-regex = "1.11.1"
-serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.140"
-
-[dev-dependencies]
-indoc = "2.0.6"
-pretty_assertions = "1.4.1"
-rstest = "0.25.0"
+[workspace]
+members = ["core"]
