+++ core/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "notion-to-markdown-core"
+version = "0.1.0"
+edition = "2021"
+
+
+[dependencies]
+async-trait = "0.1.88"
+comrak = "0.39.0"
+regex = "1.11.1"
+serde = { version = "1.0.219", features = ["derive"] }
+
+
+[dev-dependencies]
+indoc = "2.0.6"
+pretty_assertions = "1.4.1"
+rstest = "0.25.0"
+serde_json = "1.0.140"
