+++ core/src/block/embed.rs
