+++ core/src/block/equation.rs
