+++ core/src/block/image.rs
