+++ core/src/block/numbered_list_item.rs
