+++ core/src/block/pdf.rs
