+++ core/src/rich_text.rs
