+++ core/src/tests/block/bulleted_list_item_response.json
