+++ core/src/tests/block/divider_response.json
