+++ core/src/tests/block/embed_response.json
