+++ core/src/tests/block/equation_response.json
