+++ core/src/tests/block/headline2_response.json
