+++ core/src/tests/block/image_response.json
