+++ core/src/tests/block/quote_response.json
